# wireless/bcm43xxx: merge frame send to once to improve the performance