--- a/drivers/wireless/ieee80211/bcm43xxx/bcmf_sdpcm.c
+++ b/drivers/wireless/ieee80211/bcm43xxx/bcmf_sdpcm.c
@@ -391,25 +391,11 @@ int bcmf_sdpcm_sendframe(FAR struct bcmf_dev_s *priv)
                (unsigned long)sframe->header.base);
 #endif
 
-  /* Write the first 4 bytes of sdpcm header */
+  /* Write the frame data (the buffer is DMA aligned here) */
 
   ret = bcmf_transfer_bytes(sbus, true, 2, 0,
                             sframe->header.base,
-                            FIRST_WORD_SIZE);
-  if (ret != OK)
-    {
-      /* TODO handle retry count and remove frame from queue + abort TX */
-
-      wlinfo("fail send frame %d\n", ret);
-      ret = -EIO;
-      goto exit_abort;
-    }
-
-  /* Write the remaining frame data (the buffer is DMA aligned here) */
-
-  ret = bcmf_transfer_bytes(sbus, true, 2, 0,
-                            sframe->header.base + FIRST_WORD_SIZE,
-                            sframe->header.len - FIRST_WORD_SIZE);
+                            sframe->header.len);
   if (ret != OK)
     {
       /* TODO handle retry count and remove frame from queue + abort TX */